Feature: support widget instance as Frame focus part in constructor
* separate frame tests
